fix: change telegram_id from INTEGER to BIGINT to support large Telegram IDs

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, Boolean, DateTime, ForeignKey, Text, JSON, Index, func
+from sqlalchemy import Column, Integer, BigInteger, String, Float, Boolean, DateTime, ForeignKey, Text, JSON, Index, func
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from app.database import Base
@@ -7,7 +7,7 @@ class User(Base):
     __tablename__ = "users"
     
     id = Column(Integer, primary_key=True)
-    telegram_id = Column(Integer, unique=True, nullable=False, index=True)
+    telegram_id = Column(BigInteger, unique=True, nullable=False, index=True)
     username = Column(String(50))
     is_active = Column(Boolean, default=True)
     is_banned = Column(Boolean, default=False)
